STRY0010002 - Create an Update Set
We would like to capture all out work on an update set so that it would be easy
for us to move all our work from one instance to another when we complete the project.

Create an update named Loaner Vehicle Request.
b. Make sure the Application field is set to the application you created on the previous Story.
c. Click the Submit and Make Current button.
d. Validate that you are using the correct update set but going to the Application picker. 
Notice that the Update set section is now showing the Update set you created. Please see screenshot.