STRY0010006 - Load Loaner Vehicle Catalog File
The Loaner Vehicle Catalog table needs to be filled in with a list of the current
vehicle inventory of the company. Import a spreadsheet into the Loaner Vehicle Catalog table and map the fields accordingly.

Data Load New App.xlsx Download this file and load it to the Loaner Vehicle Catalog Table.
Use Import sets and Transform maps to do this.
b. Once all the records have been loaded go to the list of Loaner Vehicle Catalog records and manually add the appropriate image to each record.