STRY0010011 - Related Lists
Relationships between tables are automatically established if a field from one table is referencing a different table. We should have the ability to track various records
related to a table to add visibility and accessibility from record to another.

Navigate to the Loaner Vehicle Catalog form and add the following Related lists.
*Vehicle Trackers
*Vehicle Servicing
b. Configure the related list so it does not show the New button on both related lists.
c. Refer to the screenshot to how the related list will be positioned on the form.
d.Configure the Vehicle Tracker’s related list “list view”.
e. Configure the Vehicle Servicing’s related list “list view”.
f. Navigate to the Vehicle Tracker and add the Vehicle Servicing related list.
g. Configure the related list so it does not show the New button.
h. Configure the related list “list view”