STRY0010012 - Form UI Policies
To dynamically change the behavior of fields on a form we would need to use UI
policies. We would like to enforce mandatory fields on some fields when other fields are
set to a certain value.

a. Navigate to the Vehicle Tracker for and create UI policies for the following scenarios.
*Make Assigned to, Location, Date Requested, Expected Return Date Mandatory if Vehicle Status is Pending Release or Out on Field
*Make Actual Return Date, Return Vehicle Status and Return Notes Editable and Mandatory if Vehicle Status is Returned for Inspection.