STRY0010013b - UI Actions
We would like to have buttons for users to use to decide on an action when the
Vehicles have been returned. Users should be able to click a Return to Warehouse
button when the user sees that the vehicle is fit to be returned to the warehouse. We
should also have a button called Send to Repair if the vehicle needs to be repaired or serviced before returning to the Warehouse.

Create a Return to Warehouse button on the Vehicle Tracker form.
*Only show this button when the Vehicle Status is Returned for Inspection, and the Tracker Status is Open
*When users click this button, Set the Vehicle Status to Returned to Warehouse.
b. Create a Send to Repair button on the Vehicle Tracker form.
*Only show this button when the Vehicle Status is Returned for Inspection, and the Tracker Status is Open
*When users click this button, Set Vehicle Status to Sent for Servicing.