STRY0010016b - Loaner LVA Flow and Workflow
Once the request is submitted, we would like to automate the creation of a vehicle tracker as well as all activities associated with the request process such as approval and delivery. The flow should be triggered when the Service Catalog is submitted. A Vehicle Tracker containing information from the catalog item is automatically created. An approval to the Loaner Vehicle Approver group is attached to the RITM ticket. Once at least one approver approves, a Delivery task is then created and assigned to the Vehicle Delivery Group.  Once the delivery task is closed the RITM ticket needs to be closed.

a. Create Groups:
*Loaner Vehicle Approvers - add the approver_user role.
*Loaner Vehicle Delivery - add the itil role.
b. Assign members to the group. Note that no user should be in both groups.
c. Create a Flow
*Trigger Flow when the catalog item is submitted.
*Create a Vehicle Tracker Record and Map the following values.
--Request Number – The current request number created after the ticket is submitted.
--Assigned to – Requested for variable on the catalog item.
--Vehicle – Select the Vehicle that you need variable on the catalog item.
--Date Requested – Date Needed variable on the catalog item.
--Expected Return Date – Date to return the vehicle variable on the catalog item.
--Location – Office Location variable on the catalog item.
--Delivery Information - Delivery Information variable on the catalog item.
--Additional Comments - Additional Comments variable on the catalog item.
--Vehicle Status – Pending Release
--Tracker Status – Open
*Send an Approval to the Loaner Vehicle Approvers group
*If Approved
--Create a Catalog Task with Short Description “Loaner Vehicle ------- Delivery Task” and assign it to the Loaner Vehicle Delivery Group.
--Once the Task is closed, Set the Request Item to Closed Complete.
--Update the Vehicle Status of the Vehicle Tracker record to Out on Field
*If Rejected
--Set the Request Item to Closed Incomplete.
--Update the Vehicle Status of the Vehicle Tracker record to Returned to Warehouse and Tracker Status to Cancelled