STRY0010018 - Creating Notifications
I would need to be able to instruct users on procedures on what to do when returning the vehicles.  I would like an automated email sent by the system when the Request Item (RITM) ticket closes.  Do not add this notification on the flow. I would need to have you create a new Notification and define the appropriate conditions as stated below.

a. When to send the notification:
*When the RITM ticket state changes to Closed Complete and the Item is Request Loaner Vehicle
b. Who will receive:
*The person who opened the ticket
c. What will it contain:
*Subject: Important Notes on returning the vehicle requested on
Body:
Hi ,
The vehicle request you entered has been delivered to your location. Please take note of this ticket number and present it to the Warehouse personnel when you return the vehicle.
Thank you,