STRY0010023 - Creating a simple Portal
Create a portal that shows a list of the vehicles.

Create a portal that shows me a list of the Vehicles and all information about the vehicle. I would also need to see a widget that allows users to access the catalog item to request for a vehicle.
Note that the requirements above are the basic things that we would expect from your portal, but if you have time, please add more enhancements to your liking.